chore: initial skeleton for sdk readme generator github action
This is an initial commit for the definition of a github action
that we will eventually use to generate consistent headers for
all of our SDKs.

This commit is a mostly unmodified version of the typescript
github actions template.  In future commits it will be cleaned
up to use our licenses, code style, etc., and then we will
build out the functionality to actually templatize the sdk
readme files.