Add README header validation for SDKs
This extends the README generator so that it can do some validation on the list of headers that appear in the README for an SDK. The goal is to allow us to enforce that all of the SDKs have some level of consistency in their section headers.

It also generates some basic headers for SDK repos and some boilerplate
information about the project status.

This is just a first cut, we will probably want to refine the
list of headers and the text over time.