feat: make readme template header validation slightly more lenient (#30)
This commit relaxes the template header validation just a bit.
It will still enforce that you have all of the required headers
in the right order, but it will now tolerant cases where you have
additional headers inserted anywhere in the template.  This allows
us to add special sections in individual language repos for cases
where the language requires some additional info that may not be
relevant to the other languages.

We also add a new required `Examples` header to make sure that
we call that out clearly a little earlier in the README.

This is to conform with the latest changes to the .NET README, which
are based on feedback we got during the bug bash.