fix: improve error logging if line number not set on markdown element (#63)
In some cases it's possible for the markdown element to not have
a line number when parsing the README template, and prior to this
commit we would throw a useless error if that happens. In this
commit we change it to report line number = 1 if that happens,
which makes the rest of the error message much more useful.